% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bcc_colours}
\alias{bcc_colours}
\title{BC Cancer Brand Colours}
\format{
## `bcc_colours`
A list of length 2:
\describe{
   \item{brand_cols_only}{15 hex codes with all brand colours included}
   \item{distinct_colours}{6 hex codes with extracted distinct colours from the brand_cols_only}
}
}
\usage{
bcc_colours
}
\description{
A list of data with BC Cancer brand colour hex codes from PHSA. To add or
customize colour sets, refer to source code in the data-raw folder.
}
\keyword{datasets}
