% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcc_palettes.R
\name{bcc_palettes}
\alias{bcc_palettes}
\title{Generate a custom palette for input into scale colour functions.}
\usage{
bcc_palettes(
  name,
  n,
  all_palettes = bccpalette::bcc_colours,
  type = c("discrete", "continuous")
)
}
\arguments{
\item{name}{Specify which colour set to use from `bcc_colours`.}

\item{n}{Number of colours to extract.}

\item{all_palettes}{Specify custom colour set (default `bcc_colours`).}

\item{type}{Specify discrete or continuous data.}
}
\value{
A list of hex codes.
}
\description{
This function is called by the scale_bcc functions to extract colours from
the custom colour set `bcc_colours`.
}
