% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_bcc_c.R
\name{scale_fill_bcc_c}
\alias{scale_fill_bcc_c}
\title{Specify fill values for continuous data.}
\usage{
scale_fill_bcc_c(name)
}
\arguments{
\item{name}{Specify which colour set from `bcc_colours` to use (base package
includes `brand_cols_only` and `distinct_colours`).}
}
\value{
Takes the place of scale_fill_gradientn in ggplot to specify colours.
}
\description{
Specify fill values for continuous data.
}
