% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_bcc_d.R
\name{scale_fill_bcc_d}
\alias{scale_fill_bcc_d}
\title{Specify fill values for discrete data.}
\usage{
scale_fill_bcc_d(name, col_vec)
}
\arguments{
\item{name}{Specify which colour set from `bcc_colours` to use (base package
includes `brand_cols_only` and `distinct_colours`).}

\item{col_vec}{Specify which column in the dataset that the colours will be
used for, in this format: `data$column`. This is used to determine n number
of colours to extract from the palette.}
}
\value{
Takes the place of scale_fill_manual in ggplot to specify colours.
}
\description{
Specify fill values for discrete data.
}
